Kozzy Stability Fix & Disconnect Handling v1.06

+++ global.h
@@ -10,14 +10,14 @@ extern int MyClientSlot;
 extern QString MyName;
 extern QString HostIP;
 extern int HostPort;
-extern int m_Dps;
+extern unsigned long m_Dps;
 extern unsigned long m_Dmg;
 extern unsigned long m_Activity;
 extern int LastColor;
-extern long hitCounter;
-extern long critCounter;
-extern long condiDmg;
-extern int critChance;
+extern unsigned long hitCounter;
+extern unsigned long critCounter;
+extern unsigned long condiDmg;
+extern unsigned int critChance;
